services/translation-service/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# AI/ML Framework
langchain==0.1.5
langchain-community==0.1.5
langgraph==0.0.25
langchain-aws==0.1.0

# AWS SDK
boto3==1.34.25
botocore==1.34.25

# Vector Store
faiss-cpu==1.7.4
pinecone-client==3.0.2

# NLP
spacy==3.7.2
rapidfuzz==3.6.1
nltk==3.8.1

# Database
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.25

# Queue
aioboto3==12.3.0

# Utilities
numpy==1.26.3
tenacity==8.2.3
structlog==24.1.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0